Add AMQ in memory test resource
With this resource we can have more control over async processing
and even avoid unnecessary async processing when not needed
